test(studio): extract shared block-decorator collector (#2278)
Addresses PR review: the block-decorator traversal was duplicated in
schema-types.test.ts (collectDecoratorComponents) and
apply-decorator-components.test.ts (componentsFor), so the two guards
could drift if applyDecoratorComponents changes scope — which already
cost a double-fix when the block-only check was added. Extract the
non-recursive block-member scan into collect-block-decorator-components.ts
and have both tests use it, keeping the traversal aligned in one place.
