+++ packages/sanity-studio/src/inputs/apply-decorator-components.test.ts
@@ -8,6 +8,7 @@ import {
 import {schemaTypes as baseSchemaTypes} from '@kcvv/sanity-schemas'
 import {describe, expect, it} from 'vitest'
 import {applyDecoratorComponents} from './apply-decorator-components'
+import {collectBlockDecoratorComponents} from './collect-block-decorator-components'
 
 const Pullquote: ComponentType<BlockDecoratorProps> = () => null
 const Accent: ComponentType<BlockDecoratorProps> = () => null
@@ -118,47 +119,12 @@ describe('applyDecoratorComponents', () => {
   // accent → accent-title fields), not just the hand-rolled fixture above. If a
   // decorator `value` is renamed or dropped, this fails instead of a stale mock.
   describe('against the real @kcvv/sanity-schemas schema', () => {
-    // Collect every block decorator `component` for `value` across each
-    // top-level type's direct `field.of[]` — the reach of the helper.
-    const componentsFor = (
-      types: readonly SchemaTypeDefinition[],
-      value: string,
-    ): unknown[] => {
-      const out: unknown[] = []
-      for (const type of types) {
-        const fields = (type as {fields?: unknown[]}).fields
-        if (!Array.isArray(fields)) continue
-        for (const field of fields) {
-          const of = (field as {of?: unknown}).of
-          if (!Array.isArray(of)) continue
-          for (const member of of) {
-            // Block members only — mirror the helper's own traversal, which
-            // skips non-block members, so the collector's scope matches.
-            if (
-              !member ||
-              typeof member !== 'object' ||
-              (member as {type?: string}).type !== 'block'
-            ) {
-              continue
-            }
-            const decorators = (member as {marks?: {decorators?: unknown}})?.marks
-              ?.decorators
-            if (!Array.isArray(decorators)) continue
-            for (const d of decorators) {
-              if (d && typeof d === 'object' && (d as {value?: string}).value === value) {
-                out.push((d as {component?: unknown}).component)
-              }
-            }
-          }
-        }
-      }
-      return out
-    }
-
     it('grafts the real pullquote + accent decorators', () => {
       const grafted = applyDecoratorComponents(baseSchemaTypes, components)
-      const pullquotes = componentsFor(grafted, 'pullquote')
-      const accents = componentsFor(grafted, 'accent')
+      // Shared collector mirrors this helper's block-only traversal — see
+      // collect-block-decorator-components.ts (kept in one place, #2278).
+      const pullquotes = collectBlockDecoratorComponents(grafted, 'pullquote')
+      const accents = collectBlockDecoratorComponents(grafted, 'accent')
       expect(pullquotes.length).toBeGreaterThan(0)
       expect(accents.length).toBeGreaterThan(0)
       expect(pullquotes.every((c) => c === Pullquote)).toBe(true)
