+++ packages/sanity-studio/src/schema-types.test.ts
@@ -10,6 +10,7 @@ import {describe, expect, it} from 'vitest'
 import {applyArticleTagsInput} from './inputs/apply-article-tags-input'
 import {applyDecoratorComponents} from './inputs/apply-decorator-components'
 import {applyRespondentPicker} from './inputs/apply-respondent-picker'
+import {collectBlockDecoratorComponents} from './inputs/collect-block-decorator-components'
 
 /**
  * Guard #2 (#2278) — graft-target completeness.
@@ -45,52 +46,6 @@ const findField = (
 ): Record<string, unknown> | undefined =>
   type?.fields?.find((f) => (f as {name?: string}).name === fieldName)
 
-/**
- * Collect the `component` of every block decorator whose `value` matches,
- * scanning each top-level type's direct `field.of[]` — the exact reach of
- * `applyDecoratorComponents` (it does not recurse into nested object fields),
- * so the assertion matches what the helper can actually graft.
- */
-function collectDecoratorComponents(
-  types: readonly SchemaTypeDefinition[],
-  value: string,
-): unknown[] {
-  const out: unknown[] = []
-  for (const type of types) {
-    const fields = (type as {fields?: unknown[]}).fields
-    if (!Array.isArray(fields)) continue
-    for (const field of fields) {
-      const of = (field as {of?: unknown}).of
-      if (!Array.isArray(of)) continue
-      for (const member of of) {
-        // Only block members carry decorators the helper grafts — mirror
-        // `applyDecoratorComponents`, which skips non-block members, so the
-        // collector's scope matches the helper's reach exactly.
-        if (
-          !member ||
-          typeof member !== 'object' ||
-          (member as {type?: string}).type !== 'block'
-        ) {
-          continue
-        }
-        const decorators = (member as {marks?: {decorators?: unknown}})?.marks
-          ?.decorators
-        if (!Array.isArray(decorators)) continue
-        for (const d of decorators) {
-          if (
-            d &&
-            typeof d === 'object' &&
-            (d as {value?: string}).value === value
-          ) {
-            out.push((d as {component?: unknown}).component)
-          }
-        }
-      }
-    }
-  }
-  return out
-}
-
 const TagsSentinel: ComponentType<ArrayOfPrimitivesInputProps<string>> = () => null
 const RespondentSentinel: ComponentType<StringInputProps> = () => null
 const PullquoteSentinel: ComponentType<BlockDecoratorProps> = () => null
@@ -125,18 +80,18 @@ describe('schema-types graft targets resolve against the real schema (#2278)', (
     // Both decorators must be declared somewhere reachable in the schema — else
     // the graft is a no-op. pullquote → bio/body fields, accent → title fields.
     expect(
-      collectDecoratorComponents(baseSchemaTypes, 'pullquote').length,
+      collectBlockDecoratorComponents(baseSchemaTypes, 'pullquote').length,
     ).toBeGreaterThan(0)
     expect(
-      collectDecoratorComponents(baseSchemaTypes, 'accent').length,
+      collectBlockDecoratorComponents(baseSchemaTypes, 'accent').length,
     ).toBeGreaterThan(0)
 
     const grafted = applyDecoratorComponents(baseSchemaTypes, {
       pullquote: PullquoteSentinel,
       accent: AccentSentinel,
     })
-    const pullquotes = collectDecoratorComponents(grafted, 'pullquote')
-    const accents = collectDecoratorComponents(grafted, 'accent')
+    const pullquotes = collectBlockDecoratorComponents(grafted, 'pullquote')
+    const accents = collectBlockDecoratorComponents(grafted, 'accent')
     expect(pullquotes.length).toBeGreaterThan(0)
     expect(accents.length).toBeGreaterThan(0)
     expect(pullquotes.every((c) => c === PullquoteSentinel)).toBe(true)
